requirements.txt
asgiref==3.10.0
Django==5.2.8
django-allauth==65.13.0
django-enum==2.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
ruff==0.14.4
sqlparse==0.5.3
